feat: add option to skip creating overlay
In response to a feature request, it adds an option to create the Live USB without  any overlay.

The default option is still to create an overlay since that's the main reason yastfli was created in the first place, but it makes sense to have a switch option to create regular Live USBs without data persistence, and it is easy to implement.

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>