feat: add preserve option
These changes add a preserve feature, by which yastfli deletes the core folders that make
the drive a Live OS and writes the new content to the existing partitions, instead of
formatting them.

All under the premise that the drive has been previously made a Live OS with yastfli.

There is no guarantee that data won't be lost and it shouldn't be used on any device
storing any data that you can't risk to lose or it would be rather inconvenient or
costly to do so.

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>