feat: test partitions before directory on preserve
It doesn't matter if p2 is fat32 and directory overlay has been selected on preserve if  p1 would fail the test to use preserve, so move the later test of p1 and p2 before the former so it fails before, if it has to.

Change order of test conditions on fat32 and directory overlay on normal mode to mirror the analogue test on preserve. A cosmetic change for consistency.

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>