feat: add new clean up routine
Clean up operations are carried out at the end, when everything else has been
successful and the script hasn't been interrupted by the user.

The new routine is able to do this just the same, but it also catches errors
and interruptions by the user to carry out these operations, leaving the system
in a cleaner state even when yastfli didn't complete successfully.

Signed-off-by: Manuel Fombuena <fombuena@outlook.com>